Move regex criptocoins validator to a new package
