+++ iped-engine/src/main/java/br/gov/pf/iped/regex/cripto/Base58CheckedValidator.java
@@ -1,4 +1,4 @@
-package br.gov.pf.iped.regex;
+package br.gov.pf.iped.regex.cripto;
 
 import java.io.IOException;
 import java.security.MessageDigest;
