+++ iped-engine/src/main/java/br/gov/pf/iped/regex/cripto/DashAddressValidatorService.java
@@ -1,4 +1,4 @@
-package br.gov.pf.iped.regex;
+package br.gov.pf.iped.regex.cripto;
 
 import java.io.File;
 import java.util.Arrays;
@@ -6,15 +6,14 @@ import java.util.List;
 
 import dpf.sp.gpinf.indexer.process.task.regex.BasicAbstractRegexValidatorService;
 
-public class DogeCoinAddressValidatorService extends BasicAbstractRegexValidatorService {
+public class DashAddressValidatorService extends BasicAbstractRegexValidatorService {
 
     private static final AltcoinBase58CheckValidator validator;
     
     static {
         validator = new AltcoinBase58CheckValidator();
-        validator.setVersionForPrefix("D", (byte) 30);
-        validator.setVersionForPrefix("A", (byte) 22);
-        validator.setVersionForPrefix("9", (byte) 22);
+        validator.setVersionForPrefix("7", (byte) 16);
+        validator.setVersionForPrefix("X", (byte) 76);
     }
 
     @Override
@@ -24,7 +23,7 @@ public class DogeCoinAddressValidatorService extends BasicAbstractRegexValidator
 
     @Override
     public List<String> getRegexNames() {
-        return Arrays.asList("CRIPTOCOIN_DOGECOIN");
+        return Arrays.asList("CRIPTOCOIN_DASH");
     }
 
     @Override
