+++ iped-engine/src/main/java/br/gov/pf/iped/regex/cripto/RippleAddressValidatorService.java
@@ -1,4 +1,4 @@
-package br.gov.pf.iped.regex;
+package br.gov.pf.iped.regex.cripto;
 
 import java.io.File;
 import java.util.Arrays;
