+++ iped-engine/src/main/resources/META-INF/services/dpf.sp.gpinf.indexer.process.task.regex.RegexValidatorService
@@ -7,15 +7,15 @@ br.gov.pf.iped.regex.BoletoRegexValidatorService
 br.gov.pf.iped.regex.TituloEleitorRegexValidatorService
 br.gov.pf.iped.regex.CreditCardRegexValidatorService
 br.gov.pf.iped.regex.ScriptValidatorService
-br.gov.pf.iped.regex.BitcoinAddressValidatorService
-br.gov.pf.iped.regex.BCHAddressValidatorService
-br.gov.pf.iped.regex.EthereumAddressValidatorService
-br.gov.pf.iped.regex.RippleAddressValidatorService
+br.gov.pf.iped.regex.cripto.BitcoinAddressValidatorService
+br.gov.pf.iped.regex.cripto.BCHAddressValidatorService
+br.gov.pf.iped.regex.cripto.EthereumAddressValidatorService
+br.gov.pf.iped.regex.cripto.RippleAddressValidatorService
 br.gov.pf.iped.regex.TelefoneRegexValidatorService
 br.gov.pf.iped.regex.PlacaRegexValidatorService
 br.gov.pf.iped.regex.IPLRegexValidatorService
 br.gov.pf.iped.regex.FacebookContactValidatorService
 br.gov.pf.iped.regex.RIFRegexValidatorService
-br.gov.pf.iped.regex.DogeCoinAddressValidatorService
-br.gov.pf.iped.regex.DashAddressValidatorService
-br.gov.pf.iped.regex.LiteCoinAddressValidatorService
+br.gov.pf.iped.regex.cripto.DogeCoinAddressValidatorService
+br.gov.pf.iped.regex.cripto.DashAddressValidatorService
+br.gov.pf.iped.regex.cripto.LiteCoinAddressValidatorService
