+++ iped-engine/src/test/java/br/gov/pf/iped/regex/cripto/RippleAddressValidatorServiceTest.java
@@ -1,8 +1,10 @@
-package br.gov.pf.iped.regex;
+package br.gov.pf.iped.regex.cripto;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
+
+import br.gov.pf.iped.regex.cripto.RippleAddressValidatorService;
 
 public class RippleAddressValidatorServiceTest {
     RippleAddressValidatorService service = new RippleAddressValidatorService();
